Working on OS X Support

+++ install.sh
@@ -44,6 +44,11 @@ done
 echo "Creating dotfiles directory at $SOURCEDIR"
 mkdir -p $SOURCEDIR
 
+if [$HOME == $SOURCEDIR ]; then
+  >&2 echo "Cannot run install script from $HOME"
+  exit 1
+fi
+
 echo "Creating symlinks to $HOME from $SOURCEDIR"
 
 # Copy all files in dotfiles/home/* to $HOME/.[filename]
@@ -53,17 +58,45 @@ do
   ln $force -s $file "$HOME/.${file##*/}"
 done
 
-# Upgrade pip
-pip install -U pip
-pip install --user git+git://github.com/lukebayes/powerline
-
+# Download Powerline Fonts
 wget https://github.com/lukebayes/powerline/raw/develop/font/PowerlineSymbols.otf \
   https://github.com/lukebayes/powerline/raw/develop/font/10-powerline-symbols.conf
 
-sudo mkdir -p /usr/share/fonts
-sudo mv PowerlineSymbols.otf /usr/share/fonts/
-sudo fc-cache -vf
-sudo mv 10-powerline-symbols.conf /etc/fonts/conf.d/
+# Upgrade Python package manager
+pip install -U pip
+# Install Powerline using pip
+pip install --user git+git://github.com/lukebayes/powerline
+
+if [ ! -e $HOME/src/solarized-dir ]; then
+  # Fetch the Solarized directory theme
+  git clone https://github.com/lukebayes/dircolors-solarized.git $HOME/src/solarized-dir || true
+fi
+
+# Install the Solarized directory theme
+ln -fs $HOME/src/solarized-dir/dircolors.256dark $HOME/.dircolors
+
+echo "Preparing to install for $(uname -s)"
+
+if [ $(uname -s) == 'Darwin' ]; then
+  # TODO: Install Powerline fonts for OS X
+  echo "Installing OS X only features"
+  mv $HOME/.bashrc $HOME/.bash_profile
+fi
+
+if [ -uname -s == 'Ubuntu' ]; then
+  echo 'Installing Powerline for Linux'
+  sudo mkdir -p /usr/share/fonts
+  sudo mv PowerlineSymbols.otf /usr/share/fonts/
+  sudo fc-cache -vf
+  sudo mv 10-powerline-symbols.conf /etc/fonts/conf.d/
+
+  if [ ! -e $HOME/src/solarized ]; then
+    # Fetch the Solarized terminal theme
+    git clone https://github.com/lukebayes/gnome-terminal-colors-solarized.git $HOME/src/solarized || true
+  fi
+  # Install the Solarized terminal theme
+  $HOME/src/solarized/install.sh
+fi
 
 # Configure VIM
 if [ $skipvim = false ]; then
@@ -78,20 +111,6 @@ if [ $skipvim = false ]; then
   # echo "Installing provided VIM plugins"
   vim -c "PluginInstall!" -c "q" -c "q"
 fi
-
-if [ ! -e $HOME/src/solarized ]; then
-  # Fetch the Solarized terminal theme
-  git clone https://github.com/lukebayes/gnome-terminal-colors-solarized.git $HOME/src/solarized || true
-fi
-# Install the Solarized terminal theme
-$HOME/src/solarized/install.sh
-
-if [ ! -e $HOME/src/solarized-dir ]; then
-  # Fetch the Solarized directory theme
-  git clone https://github.com/lukebayes/dircolors-solarized.git $HOME/src/solarized-dir || true
-fi
-# Install the Solarized directory theme
-ln -fs $HOME/src/solarized-dir/dircolors.256dark $HOME/.dircolors
 
 # if [ ! -e $HOME/src/solarized-vim ]; then
 #   # Fetch the Solarized vim theme
